Renamed: FileObject to StudyObject
Introduced the new nomenclature, since the term FileObject was a little bit misleading. While it will represent a single netCDF file in the end, the term "study" is more consistent with the rest of the class names (Cell, Technique, etc.). In that sense, the in-memory representation of the netCDF file is a study object, which again can be saved to a file.